fix compose: add PGUSER and PGDATABASE env to db
Without them, a recurrent error would be displayed on the db container:

```
FATAL: role "root" does not exist
```

This is because the `pg_isready` command would not reuse the `POSTGRES_USER`
and `POSTGRES_DB` env vars. It rather uses the `PGUSER` and `PGDATABASE` env
variables, and they have to match.